totp to request.POST

+++ noxcrux_api/views/OTP.py
@@ -28,6 +28,7 @@ def get_user_totp_device(user, confirmed=None):
         ]),
 )
 class TOTPView(APIView):
+
     def get(self, request):
         device = get_user_totp_device(request.user)
         if not device:
@@ -35,25 +36,25 @@ class TOTPView(APIView):
         url = device.config_url
         return Response(url, status=status.HTTP_201_CREATED)
 
-    def post(self, request, totp_code):
+    def post(self, request):
         device = get_user_totp_device(request.user)
         if not device:
             return Response(dict(errors=['You have not setup two factor authentication.']), status=status.HTTP_400_BAD_REQUEST)
         if device.confirmed:
             return Response(dict(errors=['You already have confirmed your device.']), status=status.HTTP_400_BAD_REQUEST)
-        if device.verify_token(totp_code):
+        if 'totp_code' in request.POST and device.verify_token(request.POST.get('totp_code')):
             device.confirmed = True
             device.save()
             return Response("Device confirmed", status=status.HTTP_200_OK)
         return Response(dict(errors=dict(token=['Invalid TOTP Token.'])), status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, totp_code):
+    def delete(self, request):
         device = get_user_totp_device(request.user)
         if not device:
             return Response(dict(errors=['You have not setup two factor authentication.']), status=status.HTTP_400_BAD_REQUEST)
         if not device.confirmed:
             return Response(dict(errors=['Please confirm your device first.']), status=status.HTTP_400_BAD_REQUEST)
-        if device.verify_token(totp_code):
+        if 'totp_code' in request.POST and device.verify_token(request.POST.get('totp_code')):
             device.delete()
             return Response("Device deleted and 2FA deactivated", status=status.HTTP_204_NO_CONTENT)
         return Response(dict(errors=dict(token=['Invalid TOTP Token.'])), status=status.HTTP_400_BAD_REQUEST)
